mkimage.sh: fix sfdisk usage for multiple arches.
In the case where we have two partitions, boot and root, the
root partition needs to start at sector (2048 + bootfs_size_in_sectors),
not at $BOOT_FSSIZE.
Using $BOOT_FSSIZE as the start tells sfdisk to try to start the
/ partition in the middle of /boot, which it refuses to do.
Leave start blank, and sfdisk uses the first unallocated sector as
the start of /, which happens to be sector 2048 + bootfs_size_in_sectors.

I suspect this fixes #81.