mknet.sh: Massive Cleanup
Among other things this commit refactors mknet.sh to use the lib.sh
functions for executing code in the chroot and generally handling the
potential for future builds to generate netboot tarballs for ARM
platforms.  ARM tarballs will require syslinux to be built for the
target architecture, but it should largely just be a drop in option.