Fixed WPA configuration and f2fs installation,
The void-installer wifi configuration never worked for me. Whenever
I installed void, I always had to configure wpa_supplicant manually.
The reason is that the void-installer uses dhcpcd to hook into
wpa_supplicant, but the void-installer doesn't copy the wpa_supplicant hook
from /usr/share/dhcpcd/hooks into /usr/libexec/dhcpcd-hooks. After I modified
the installer to do this during wifi configuration, it worked! I also set it
up to copy the wpa_supplicant hook into /usr/libexec/dhcpcd-hooks in the
installation directory. So WI-FI should work out of the box now!

Another issue I had with the void installer was that I could never get f2fs
to work on void. This is because the kernel is stored in /boot, but the EFI
partition is mounted at /boot/efi. So if f2fs is used for the root filesystem,
then grub will return an unknown filesystem error since it doesn't support
f2fs.

To fix this, I modified the installer to allow the EFI system partition to
optionally be mounted at /boot instead of /boot/efi. After doing this, I
installed void on a flash drive with f2fs for the root partition and the
EFI system partition mounted at /boot and it worked fine! I also added the
-f flag to the mkfs.f2fs command in the create_filesystems() function. Before,
the installer wouldn't overwrite an existing f2fs partition, even if you
specified that you wanted a new filesystem to be created.

I also added support for nilfs2 in the installer. However, the installer
script doesn't add nilfs-utils into the live image by default, so you
have to specify it when running the installer script via the -p option.
Based on tests I ran, grub doesn't detect kernels on nilfs2 partitions,
even though the GRUB manual says that GRUB supports nilfs2.
So like f2fs, if you want to use nilfs2, the EFI system
partition has to be mounted at /boot.

Another issue I ran into was that the installer.sh script doesn't
automatically include dialog (the program the installer uses for the TUI)
in the live image. It has to be specified via the -p option.

However, I ran into one issue which I haven't fixed yet. When installing void
on a flash drive, everything went smoothly, but when I rebooted, my UEFI boot
manager failed to load the grub EFI binary. I had to reconfigure it in the
UEFI settings, after which it booted up fine. I ran into pretty much the same issue with
the official void installer ISO (the one on the website). So I don't think this bug was
caused by the changes I made to the installer.