mkimage.sh.in: explicitly set 'PermitRootLogin yes' in sshd_config
So far, image generation relied on the config shipped in the openssh
package to ensure root login with password, but as this might change in
the future, explicitly apply and document this requirement for embedded
images here instead.

The exception for GCP (disable root login) should still work as
intended.