packer: Add AWS AMI
This commit adds a packer configuration to build an AWS AMI.  The AMI
works in that it is a bootable void image that if it could be imported
to AWS it would be bootable and would work.  Unfortunately, AWS is
doing really naive string parsing on files in /boot rather than
checking that the contents are actually what they claim to be so the
import task fails.

NOTICE: AWS is the only cloud that this has been observed happening
on.  We can run as a vendor supported distro in GCP and in Azure
clouds, and so if you need Void on an enterprise cloud those are the
providers to look at.  If you're interested in a hobbyist cloud, Void
runs critical infrastructure in DigitalOcean and it is a fully
supported environment.

We could probably masquerade the kernel as something that would pass
the string parsing, because it appears that this parsing layer exists
only in the vm import service and not actually in any of the
functional parts of EC2, but that would require using a different
mechanism to install Void, and the ROI on supporting a single cloud in
this way just isn't there.  This import mechanism also appears to only
work on amd64 architectures, so it can't even support the newer
graviton instance types.