installer.sh.in: Fixing DONE variables for set_useraccount and source
The issue with the various USER_DONE variables is that they are reset if
the installation process was aborted and then restarted. The configuration
were still there but the user wasn't created because of the DONE variables.
Instead of checking the DONE variables now the actual user settings are
checked for existence.

Another bug with the SOURCE_DONE is also fixed. Previous the local image
was installed instead of the net packages if installation process was
aborted and later continued by the user.

A third fix touches the test_network function. Instead of first rm the
downloaded file and then execute the xbps-uhelper command the execution
order was switched. This ensures that the downloaded file isn't there
right after the download process. This fix keeps home direcotries clean.