installer.sh.in: Add POSTINSTALL to execute script for additional steps after installation
For more automation in the install process the POSTINSTALL settings adds
the ability to execute a script after the base installation process to
configure the installation before reboot.

This gives the ability to automate the installation process and install
software, wm's or desktop environments flawlessly.

The intention of this functionality is to give users the ability to
configure the void installation for their needs at installation time.
Could also be a good place to extend the installer functionality with
custom scripts, e.g. to set up a working development VM with all needed
packages and environment variables.