build-x86-images: Fix passing of additional arguments to mklive
Mklive is called with "$@", to be able to pass additionnal arguments to it.
However, since mklive is called from within a function, the value of "$@" is
the function's argument.

This is fixed by adding "$@" to the function call and using shift after its
first argument (the image flavour) is used.