mklive.sh: check arch validity after the default is set
It's possible to run mklive.sh without the -a argument in which case,
the architecture is supposed to be set to the current one using
xbps-uhelper. This means we have to check whether the architecture is
valid after this default has been set, otherwise we get an empty string
which is not a valid architecture.