installer.sh: fix USER*_DONE, validate user account
also make the description of the USERNAME (comment/GECOS) more
informative

from dateiexplorer:
The issue with the various USER_DONE variables is that they are reset if
the installation process was aborted and then restarted. The configuration
were still there but the user wasn't created because of the DONE variables.
Instead of checking the DONE variables now the actual user settings are
checked for existence.

Co-authored-by: dateiexplorer <justus.roederer@dateiexplorer.de>