mkimage: reduce default total image size to 512 + 256 MiB
because growpart is now run at boot, we can make the base image smaller.
This does not provide any benefit when compressed (the image is still
81M compressed with xz 9), but this should be much faster to write to
the target SD card/USB drive for SBCs, where the throughput is rather
slow. Growpart will inflate the root partition to a reasonable size for
the user. Users who want more flexibility in partitioning should be
doing a chroot/PLATFORMFS install.