installer.sh: add post-installation menu for enabling services
chose to do it *post* install, just before syncing and unmounting,
because it's the easiest place to know exactly what services will be
available and which will be enabled in the target system in the mess
that is the el-cheapo installer.

filters out some important services that users shouldn't be messing with
at this point in the install, like some agettys, udevd, and sulogin.